Serve local data from controller

+++ update-app-remotely.sh
@@ -1,14 +1,20 @@
 #!/bin/bash
 
 if [[ $1 == "" ]]; then
-  echo "Usage: ./update-app-remotely.sh <controller_ip>";
+  echo "Usage: ./update-app-remotely.sh <controller_ip> [<source directory>]";
   exit;
 fi;
 
+if [[ $2 == "" ]]; then
+  source_dir = './frontend';
+  git clone https://github.com/m4mcontroller/frontend $source_dir;
+else
+  source_dir=$2;
+fi;
+
 controller_host=$1;
- 
-git clone https://github.com/m4mcontroller/frontend;
-cd frontend;
+
+cd $source_dir;
 
 npm install;
 MODE=lite npm run build -- --environment=production;
@@ -16,5 +22,6 @@ MODE=lite npm run build -- --environment=production;
 scp -r dist pi@"$controller_host":/tmp;
 ssh pi@"$controller_host" 'sudo rm -rf /usr/html; sudo mv /tmp/dist /usr/html/;'
 
-cd ..;
-rm -rf frontend;
+if [[ $2 == "" ]]; then
+  rm -rf frontend;
+fi;
